Added deletion confirmation, started collection details component

+++ MyCollectionClient/src/app/collection-list/collection-list.component.ts
@@ -1,5 +1,5 @@
-import { Component, OnInit } from '@angular/core';
-import { Collection } from '../services/collections/collection';
+import {Component, OnInit} from '@angular/core';
+import {Collection} from '../services/collections/collection';
 import {CollectionService} from '../services/collections/collection.service';
 
 @Component({
@@ -9,9 +9,12 @@ import {CollectionService} from '../services/collections/collection.service';
 })
 export class CollectionListComponent implements OnInit {
 
+  openedCollection: Collection;
+
   collections: Collection[];
 
-  constructor(private collectionService: CollectionService) { }
+  constructor(private collectionService: CollectionService) {
+  }
 
   ngOnInit() {
     this.getCollections();
@@ -22,7 +25,17 @@ export class CollectionListComponent implements OnInit {
   }
 
   delete(collection: Collection): void {
-    this.collections = this.collections.filter(c => c !== collection);
-    this.collectionService.deleteCollection(collection).subscribe();
+    if (confirm('Do you really want to delete the collection? This will remove the index but does not delete the media files.')) {
+      this.collections = this.collections.filter(c => c !== collection);
+      this.collectionService.deleteCollection(collection).subscribe();
+    }
+  }
+
+  toggleDetails(collection: Collection): void {
+    if (this.openedCollection === collection) {
+      this.openedCollection = undefined;
+    } else {
+      this.openedCollection = collection;
+    }
   }
 }
